Added a state handler for the enemy.
The state handler will check how far the enemy is from the player. Depending on the distance, its state will change to either patrolling, chasing, or attacking.

Also modified the EnemyPatrol class to derive from the EnemyStateHandler class, and will only function when the state of the enemy is "Patrol".